+++ src/unlp/ttpsInfoPoolCBR/dao/IGenericDao.java
@@ -11,7 +11,7 @@ public interface IGenericDao<M extends AbstractEntity,VO extends AbstractEntity>
 
     public VO guardar(VO objetoVO)throws Exception;
     public VO modificar(VO objetoVO)throws Exception;
-    public void borrar(Long idObjetoVO) throws Exception;
+    public void borrar(Integer idObjetoVO) throws Exception;
 
     public List<VO> listar()throws Exception;
     public VO encontrar(Integer id) throws Exception;
